drobne zmiany

+++ src/main/java/pl/ryzykowski/arduino/arduinotemperaturews/repository/TemperatureRepository.java
@@ -7,6 +7,6 @@ import java.time.LocalDateTime;
 
 public interface TemperatureRepository extends JpaRepository<Temperature, Long> {
 
-    void removeByLocationIdAndTimestampLessThan(int locationId, LocalDateTime timestamp);
+    void removeByLocationIdAndTimestampLessThan(long locationId, LocalDateTime timestamp);
 
 }
